requirements_3_11_1_mac.txt
Package                   Version
------------------------- -----------
aiofiles                  24.1.0
aiohappyeyeballs          2.5.0
aiohttp                   3.11.13
aiosignal                 1.3.2
altgraph                  0.17.4
annotated-types           0.7.0
anyio                     4.8.0
attrs                     25.1.0
bidict                    0.23.1
bottle                    0.13.2
certifi                   2025.1.31
cffi                      1.17.1
charset-normalizer        3.4.1
click                     8.1.8
clr_loader                0.2.7.post0
colorama                  0.4.6
docutils                  0.21.2
fastapi                   0.115.11
frozenlist                1.5.0
h11                       0.14.0
httpcore                  1.0.7
httptools                 0.6.4
httpx                     0.28.1
idna                      3.10
ifaddr                    0.2.0
iniconfig                 2.0.0
itsdangerous              2.2.0
Jinja2                    3.1.6
macholib                  1.16.3
markdown2                 2.5.3
MarkupSafe                3.0.2
multidict                 6.1.0
nicegui                   2.12.1
numpy                     2.2.3
orjson                    3.10.15
packaging                 24.2
pandas                    2.2.3
pefile                    2023.2.7
pip                       24.2
pluggy                    1.5.0
propcache                 0.3.0
proxy_tools               0.1.0
pscript                   0.7.7
psycopg2-binary           2.9.10
pycparser                 2.22
pydantic                  2.10.6
pydantic_core             2.27.2
Pygments                  2.19.1
pyinstaller               6.12.0
pyinstaller-hooks-contrib 2025.1
pyobjc-core               11.0
pyobjc-framework-Cocoa    11.0
pyobjc-framework-Quartz   11.0
pyobjc-framework-Security 11.0
pyobjc-framework-WebKit   11.0
pytest                    8.3.5
python-dateutil           2.9.0.post0
python-dotenv             1.0.1
python-engineio           4.11.2
python-multipart          0.0.20
python-socketio           5.12.1
pythonnet                 3.0.5
pytz                      2025.1
pywebview                 5.4
pywin32-ctypes            0.2.3
PyYAML                    6.0.2
requests                  2.32.3
setuptools                76.0.0
simple-websocket          1.1.0
six                       1.17.0
sniffio                   1.3.1
starlette                 0.46.1
typing_extensions         4.12.2
tzdata                    2025.1
urllib3                   2.3.0
uvicorn                   0.34.0
uvloop                    0.21.0
vbuild                    0.8.2
watchfiles                1.0.4
websockets                15.0.1
wsproto                   1.2.0
yarl                      1.18.3
